\name{Atom}
\alias{Atom}

\docType{data}
\title{
Atom from FCPS
}
\description{
Two nested spheres with different variances that are not linear not separable.
}
\usage{data("Atom")}

\details{
Size 800, Dimensions 3, stored in \code{Atom$Data}

Classes 2, stored in \code{Atom$Cls}
}

\references{
Ultsch, A.: Clustering with SOM: U*C, In Proc. Workshop on Self-Organizing Maps, Paris, France, (2005) , pp. 75-82
}
\examples{
data(Atom)
str(Atom)
}
\keyword{datasets,Atom,FCPS}