\name{GolfBall}
\alias{GolfBall}

\docType{data}
\title{
GolfBall from FCPS
}
\description{
no clusters at all
}
\usage{data("GolfBall")}

\details{
Size 4002, Dimensions 3, stored in \code{GolfBall$Data}

Classes 1, stored in \code{GolfBall$Cls}
}

\references{
Ultsch, A.: Clustering with SOM: U*C, In Proc. Workshop on Self-Organizing Maps, Paris, France, (2005) , pp. 75-82
}
\examples{
data(GolfBall)
str(GolfBall)
}
\keyword{GolfBall}
\keyword{datasets}
\keyword{FCPS}