\name{Lsun}
\alias{Lsun}

\docType{data}
\title{
Lsun from FCPS
}
\description{
different variances and inter cluster distances
}
\usage{data("Lsun")}

\details{
Size 400, Dimensions 2, stored in \code{Lsun$Data}

Classes 3, stored in \code{Lsun$Cls}
}

\references{
Ultsch, A.: Clustering with SOM: U*C, In Proc. Workshop on Self-Organizing Maps, Paris, France, (2005) , pp. 75-82
}
\examples{
data(Lsun)
str(Lsun)
}
\keyword{datasets,Lsun,FCPS}