\name{MinSpanTree}
\Rdversion{1.1}
\alias{MinSpanTree}
\title{
 Zeichnet einen 2 dimensionalen minimal spanning Tree 
}
\description{
Clustering with Minimum Spanning Tree
}
\usage{
MinSpanTree(DataOrDistance, isDistance)
}
\arguments{
\item{DataOrDistance}{Der Datensatz oder die Distanzmatrix}
\item{isDistance}{Setze gleich True falls die Eingabe Distanzen sind}
}

\value{TR                    an object of class spantree which is a list with two vectors,}

