\name{NeuralGas}
\alias{NeuralGas}
\title{   liefert eine Klassenzuweisung}
\arguments{
\item{Data[1:n,1:d]}{Der Datensatz}
\item{FestgesetzteClustAnz}{in soviele Cluster werden die daten eingeteilt}
}
\value{Cls[1:n]                Clusterung der Daten}
\description{
   liefert eine Klassenzuweisung}
\author{
MT 06/2015}
