\name{Target}
\alias{Target}

\docType{data}
\title{
Target from FCPS
}
\description{
outliers
}
\usage{data("Target")}

\details{
Size 770, Dimensions 2, stored in \code{Target$Data}

Classes 6, stored in \code{Target$Cls}
}

\references{
Ultsch, A.: Clustering with SOM: U*C, In Proc. Workshop on Self-Organizing Maps, Paris, France, (2005) , pp. 75-82
}
\examples{
data(Target)
str(Target)
}
\keyword{Target}
\keyword{datasets}
\keyword{FCPS}