\name{Tetra}
\alias{Tetra}

\docType{data}
\title{
Tetra form FCPS
}
\description{
almost touching clusters
}
\usage{data("Tetra")}

\details{
Size 400, Dimensions 3, stored in \code{Tetra$Data}

Classes 4, stored in \code{Tetra$Cls}
}

\references{
Ultsch, A.: Clustering with SOM: U*C, In Proc. Workshop on Self-Organizing Maps, Paris, France, (2005) , pp. 75-82
}
\examples{
data(Tetra)
str(Tetra)
}
\keyword{datasets,Tetra,FCPS}
