\name{TwoDiamonds}
\alias{TwoDiamonds}

\docType{data}
\title{
TwoDiamonds form FCPS
}
\description{
cluster border defined by density
}
\usage{data("TwoDiamonds")}

\details{
Size 800, Dimensions 2, stored in \code{TwoDiamonds$Data}

Classes 2, stored in \code{TwoDiamonds$Cls}
}

\references{
Ultsch, A.: Clustering with SOM: U*C, In Proc. Workshop on Self-Organizing Maps, Paris, France, (2005) , pp. 75-82
}
\examples{
data(TwoDiamonds)
str(TwoDiamonds)
}
\keyword{datasets,TwoDiamonds,FCPS}
