\name{WingNut}
\alias{WingNut}

\docType{data}
\title{
WingNut of [Ultsch, 2005]
}
\description{
density vs. distance. Detailed description of dataset and its clustering challenge is provided in [Thrun/Ultsch, 2020].
}
\usage{data("WingNut")}

\details{
Size 1016, Dimensions 2, stored in \code{WingNut$Data}

Classes 2, stored in \code{WingNut$Cls}
}

\references{
[Ultsch, 2005]  Ultsch, A.: Clustering wih SOM: U* C, Proc. Proceedings of the 5th Workshop on Self-Organizing Maps, Vol. 2, pp. 75-82, 2005.
}
\examples{
data(WingNut)
str(WingNut)
}
\keyword{WingNut}
\keyword{datasets}
\keyword{FCPS}
