\name{Chainlink}
\alias{Chainlink}

\docType{data}
\title{
Chainlink of [Ultsch et al., 1994; Ultsch, 1995]. 
}
\description{
Two chains of rings. Detailed description of dataset and its clustering challenge is provided in [Thrun/Ultsch, 2020].
}
\usage{data("Chainlink")}

\details{
Size 1000, Dimensions 3, stored in \code{Chainlink$Data}

Classes 2, stored in \code{Chainlink$Cls}
}

\references{
[Ultsch et al., 1994]  Ultsch, A., Guimaraes, G., Korus, D., & Li, H.: Knowledge extraction from artificial neural networks and applications, Parallele Datenverarbeitung mit dem Transputer, (pp. 148-162), Springer, 1994.

[Ultsch, 1995]  Ultsch, A.: Self organizing neural networks perform different from statistical k-means clustering, Proc. Society for Information and Classification (GFKL), Vol. 1995, Basel 8th-10th March 1995.


[Thrun/Ultsch, 2020]  Thrun, M. C., Ultsch, A.: Clustering Benchmark Datasets Exploiting the Fundamental Clustering Problems, Data in Brief, 2020.

}
\examples{
data(Chainlink)
str(Chainlink)
}
\keyword{Chainlink}
\keyword{datasets}
\keyword{FCPS}