\name{Chainlink}
\alias{Chainlink}

\docType{data}
\title{
Chainlink from FCPS
}
\description{
linear not separable
}
\usage{data("Chainlink")}

\details{
Size 1000, Dimensions 3, stored in \code{Chainlink$Data}

Classes 2, stored in \code{Chainlink$Cls}
}

\references{
Ultsch, A.: Clustering with SOM: U*C, In Proc. Workshop on Self-Organizing Maps, Paris, France, (2005) , pp. 75-82
}
\examples{
data(Chainlink)
str(Chainlink)
}
\keyword{Chainlink}
\keyword{datasets}
\keyword{FCPS}