\name{EngyTime}
\alias{EngyTime}

\docType{data}
\title{
EngyTime from FCPS
}
\description{
Gaussian mixture
}
\usage{data("EngyTime")}

\details{
Size 4096, Dimensions 2, stored in \code{EngyTime$Data}

Classes 2, stored in \code{EngyTime$Cls}
}

\references{
Ultsch, A.: Clustering with SOM: U*C, In Proc. Workshop on Self-Organizing Maps, Paris, France, (2005) , pp. 75-82
}
\examples{
data(EngyTime)
str(EngyTime)
}
\keyword{EngyTime}
\keyword{datasets}
\keyword{FCPS}