\name{FCPS-package}
\alias{FCPS-package}
\alias{ClusteringAlgorithms}

\docType{package}
\title{
FCPC Data Sets with Common Clustering Algorithms
}
\description{
The Fundamental Clustering Problems Suite (FCPS) offers a variety of clustering problems any algorithm shall be able to handle when facing real world data. FCPS serves as an elementary benchmark for clustering algorithms.

FCPS consists of data sets with known a priori classifications that are to be reproduced by the algorithm. All data sets are intentionally created to be simple and might be visualized in two or three dimensions. Each data sets represents a certain problem that is solved by known clustering algorithms with varying success. This is done in order to reveal benefits and shortcomings of algorithms in question. Standard clustering methods, e.g. single-linkage, ward und k-means, are not able to solve all FCPS problems satisfactorily.

FCPS is supposed to be used in scientific works for free, as long as it is quotet as follows:
}

\author{
 Michael Thrun, Felix Pape, Alfred Ultsch
 
 Maintainer: Michael Thrun <mthrun@mathematik.uni-marburg.de>
}
\references{
Ultsch, A.: Clustering with SOM: U*C, In Proc. Workshop on Self-Organizing Maps, Paris, France, (2005) , pp. 75-82
}

\keyword{FCPS}
\keyword{data set}
\keyword{benchmarking}
\keyword{clustering}
\keyword{cluster}