\name{GolfBall}
\alias{GolfBall}

\docType{data}
\title{
GolfBall introduced in [Ultsch, 2005]
}
\description{
No clusters at all. Detailed description of dataset and its clustering challenge is provided in [Thrun/Ultsch, 2020].
}
\usage{data("GolfBall")}

\details{
Size 4002, Dimensions 3, stored in \code{GolfBall$Data}

Classes 1, stored in \code{GolfBall$Cls}
}

\references{
[Ultsch, 2005]  Ultsch, A.: Clustering wih SOM: U* C, Proc. Proceedings of the 5th Workshop on Self-Organizing Maps, Vol. 2, pp. 75-82, 2005.

[Thrun/Ultsch, 2020]  Thrun, M. C., & Ultsch, A.: Clustering Benchmark Datasets Exploiting the Fundamental Clustering Problems, Data in Brief, Vol. in press, pp. 105501, \doi{10.1016/j.dib.2020.105501}, 2020.

}
\examples{
data(GolfBall)
str(GolfBall)
}
\keyword{GolfBall}
\keyword{datasets}
\keyword{FCPS}