\name{Hepta}
\alias{Hepta}

\docType{data}
\title{
Hepta introduced in [Ultsch, 2003]
}
\description{
Clearly defined clusters, different variances. Detailed description of dataset and its clustering challenge is provided in [Thrun/Ultsch, 2020].
}
\usage{data("Hepta")}

\details{
Size 212, Dimensions 3, stored in \code{Hepta$Data}

Classes 7, stored in \code{Hepta$Cls}

}

\references{
[Ultsch, 2003]  Ultsch, A.: Maps for the visualization of high-dimensional data spaces, Proc. Workshop on Self organizing Maps (WSOM), pp. 225-230, Kyushu, Japan, 2003.

[Thrun/Ultsch, 2020]  Thrun, M. C., & Ultsch, A.: Clustering Benchmark Datasets Exploiting the Fundamental Clustering Problems, Data in Brief, Vol. in press, pp. 105501, \doi{10.1016/j.dib.2020.105501}, 2020.
}
\examples{
data(Hepta)
str(Hepta)
}
\keyword{Hepta}
\keyword{datasets}
\keyword{FCPS}
