\name{Hepta}
\alias{Hepta}

\docType{data}
\title{
Hepta form FCPS
}
\description{
clearly defined clusters, different variances
}
\usage{data("Hepta")}

\details{
Size 212, Dimensions 3, stored in \code{Hepta$Data}

Classes 7, stored in \code{Hepta$Cls}
}

\references{
Ultsch, A.: Clustering with SOM: U*C, In Proc. Workshop on Self-Organizing Maps, Paris, France, (2005) , pp. 75-82
}
\examples{
data(Hepta)
str(Hepta)
}
\keyword{datasets,Hepta,FCPS}
