\name{Target}
\alias{Target}

\docType{data}
\title{
Target introduced in [Ultsch, 2005].
}
\description{
Detailed description of dataset and its clustering challenge of outliers is provided in [Thrun/Ultsch, 2020]
}
\usage{data("Target")}

\details{
Size 770, Dimensions 2, stored in \code{Target$Data}

Classes 6, stored in \code{Target$Cls}
}

\references{
[Ultsch, 2005]  Ultsch, A.: U* C: Self-organized Clustering with Emergent Feature Maps, Proc. Lernen, Wissensentdeckung und Adaptivitaet (LWA/FGML), pp. 240-244, Saarbruecken, Germany, 2005.

[Thrun/Ultsch, 2020]  Thrun, M. C., & Ultsch, A.: Clustering Benchmark Datasets Exploiting the Fundamental Clustering Problems, Data in Brief, Vol. 30(C), pp. 105501, \doi{10.1016/j.dib.2020.105501}, 2020.
}
\examples{
data(Target)
str(Target)
}
\keyword{Target}
\keyword{datasets}
\keyword{FCPS}