\name{Tetra}
\alias{Tetra}

\docType{data}
\title{
Tetra introduced in [Ultsch, 1993]
}
\description{
Almost touching clusters. Detailed description of dataset and its clustering challenge is provided in [Thrun/Ultsch, 2020].
}
\usage{data("Tetra")}

\details{
Size 400, Dimensions 3, stored in \code{Tetra$Data}

Classes 4, stored in \code{Tetra$Cls}
}

\references{
[Ultsch, 1993]  Ultsch, A.: Self-organizing neural networks for visualisation and classification, Information and classification, (pp. 307-313), Springer, 1993.

[Thrun/Ultsch, 2020]  Thrun, M. C., & Ultsch, A.: Clustering Benchmark Datasets Exploiting the Fundamental Clustering Problems, Data in Brief, Vol. 30(C), pp. 105501, \doi{10.1016/j.dib.2020.105501}, 2020.
}
\examples{
data(Tetra)
str(Tetra)
}
\keyword{Tetra}
\keyword{datasets}
\keyword{FCPS}
