\name{TwoDiamonds}
\alias{TwoDiamonds}

\docType{data}
\title{
TwoDiamonds introduced in  [Ultsch, 2003a, 2003b]
}
\description{
Cluster border defined by density. Detailed description of dataset and its clustering challenge is provided in [Thrun/Ultsch, 2020].
}
\usage{data("TwoDiamonds")}

\details{
Size 800, Dimensions 2, stored in \code{TwoDiamonds$Data}

Classes 2, stored in \code{TwoDiamonds$Cls}
}

\references{
[Ultsch, 2003a]  Ultsch, A.Optimal density estimation in data containing clusters of unknown structure, technical report, Vol. 34,University of Marburg, Department of Mathematics and Computer Science, 2003.

[Ultsch, 2003b]  Ultsch, A.: U*-matrix: a tool to visualize clusters in high dimensional data, Fachbereich Mathematik und Informatik, 2003.

[Thrun/Ultsch, 2020]  Thrun, M. C., & Ultsch, A.: Clustering Benchmark Datasets Exploiting the Fundamental Clustering Problems, Data in Brief, Vol. in press, pp. 105501, \doi{10.1016/j.dib.2020.105501}, 2020.

}
\examples{
data(TwoDiamonds)
str(TwoDiamonds)
}
\keyword{TwoDiamonds}
\keyword{datasets}
\keyword{FCPS}
