\name{WingNut}
\alias{WingNut}

\docType{data}
\title{
WingNut form FCPS
}
\description{
density vs. distance
}
\usage{data("WingNut")}

\details{
Size 1016, Dimensions 2, stored in \code{WingNut$Data}

Classes 2, stored in \code{WingNut$Cls}
}

\references{
Ultsch, A.: Clustering with SOM: U*C, In Proc. Workshop on Self-Organizing Maps, Paris, France, (2005) , pp. 75-82
}
\examples{
data(WingNut)
str(WingNut)
}
\keyword{WingNut}
\keyword{datasets}
\keyword{FCPS}
