\name{WingNut}
\alias{WingNut}

\docType{data}
\title{
WingNut introduced in [Ultsch, 2005]
}
\description{
Density vs. distance. Detailed description of dataset and its clustering challenge is provided in [Thrun/Ultsch, 2020].
}
\usage{data("WingNut")}

\details{
Size 1016, Dimensions 2, stored in \code{WingNut$Data}

Classes 2, stored in \code{WingNut$Cls}
}

\references{
[Ultsch, 2005]  Ultsch, A.: Clustering wih SOM: U* C, Proc. Proceedings of the 5th Workshop on Self-Organizing Maps, Vol. 2, pp. 75-82, 2005.

[Thrun/Ultsch, 2020]  Thrun, M. C., & Ultsch, A.: Clustering Benchmark Datasets Exploiting the Fundamental Clustering Problems, Data in Brief, Vol. in press, pp. 105501, \doi{10.1016/j.dib.2020.105501}, 2020.
}
\examples{
data(WingNut)
str(WingNut)
}
\keyword{WingNut}
\keyword{datasets}
\keyword{FCPS}
